refactor: remove circle engine and move function getNearestPointBetweenPointAndEllipse into math file #WIK-15149
