+++ .changeset/violet-olives-allow.md
@@ -1,0 +1,7 @@
+---
+'@plait/draw': patch
+---
+
+remove circle engine
+
+move function getNearestPointBetweenPointAndEllipse into math file
