+++ packages/draw/src/engines/flowchart/or.ts
@@ -2,7 +2,7 @@ import { PlaitBoard, RectangleClient } from '@plait/core';
 import { PlaitGeometry, ShapeEngine } from '../../interfaces';
 import { Options } from 'roughjs/bin/core';
 import { getTextRectangle } from '../../utils';
-import { createEllipseEngine } from '../basic-shapes/circle';
+import { createEllipseEngine } from '../basic-shapes/ellipse';
 
 export const OrEngine: ShapeEngine = createEllipseEngine({
     draw(board: PlaitBoard, rectangle: RectangleClient, options: Options) {
